feat: consolidate enums, clean documentation, and prepare for open source release
Major Improvements:
- Consolidate duplicate EvidenceType enums into shared types.py for consistency
- Remove personal information from all documentation for open source readiness
- Move CLAUDE.md to private .claude/ folder (outside git repo)
- Update SECURITY.md to use GitHub Security Advisories instead of personal email
- Clean up obsolete documentation files (DEVELOPMENT.md, TESTING_STATUS.md)
- Fix enum comparison issues in tests with shared type definitions

Documentation Updates:
- Update CONTRIBUTING.md with correct repository name and contact methods
- Fix docs/README.md references to match actual available files
- Remove personal email addresses from CODE_OF_CONDUCT.md and SECURITY.md
- Update supported versions in SECURITY.md to include 0.2.x

Technical Fixes:
- Fix test failures caused by duplicate EvidenceType enum definitions
- Import shared enums from types.py in all modules
- Ensure all tests use consistent enum references
- Update security baseline for clean repository state

Status: Ready for open source release - all personal information removed
