+++ .secrets.baseline
@@ -127,5 +127,5 @@
     }
   ],
   "results": {},
-  "generated_at": "2025-08-26T14:25:07Z"
+  "generated_at": "2025-08-28T23:06:24Z"
 }
