--- CLAUDE.md
@@ -1,138 +1,0 @@
-# CLAUDE.md - Harmonic Analysis Python Library
-
-This file provides guidance to Claude Code when working with this harmonic analysis library.
-
-## Project Overview
-
-This is a comprehensive harmonic analysis library ported from TypeScript, designed to provide sophisticated harmonic and
-modal analysis with multiple interpretations. The library excels at modal analysis while handling functional harmony,
-chromatic harmony, and complex theoretical relationships.
-
-**Status: Production Ready with Calibrated Confidence Scoring**
-
-- ✅ 96% behavioral parity with TypeScript implementation achieved
-- ✅ Core functionality fully operational
-- ✅ Comprehensive test suite with 427 sophisticated test cases
-- ✅ **Confidence calibration completed** - theoretically sound cadence-specific scoring
-- ✅ Test performance: Modal (56%+), Functional (50%+) passing validation targets
-
-## Quick Start
-
-```python
-from harmonic_analysis import analyze_progression_multiple
-
-# Simple analysis
-result = await analyze_progression_multiple(['C', 'F', 'G', 'C'])
-print(f"Primary: {result.primary_analysis.analysis}")
-print(f"Confidence: {result.primary_analysis.confidence:.2f}")
-```
-
-## Documentation
-
-- **[API Guide](docs/API_GUIDE.md)** - Complete API usage examples and integration patterns
-- **[Architecture](docs/ARCHITECTURE.md)** - Technical background, evidence framework, and implementation details
-- **[Testing](docs/TESTING.md)** - Test framework, edge cases, and validation approach
-- **[Development](docs/DEVELOPMENT.md)** - Development commands, code quality standards, and maintenance
-- **[Troubleshooting](docs/TROUBLESHOOTING.md)** - Common issues, debugging, and advanced usage patterns
-- **[Confidence Calibration](docs/CONFIDENCE_CALIBRATION.md)** - Detailed confidence scoring implementation
-- **[Migration History](docs/MIGRATION_HISTORY.md)** - TypeScript to Python migration and production readiness
-
-## Key Architecture Concepts
-
-### Parent Key + Local Tonic Approach
-
-**CRITICAL**: All modal analysis uses consistent Parent Key Signature + Local Tonic approach:
-
-- **Parent Key Signature**: Underlying scale (e.g., C major, no sharps/flats)
-- **Local Tonic**: Note functioning as tonal center (e.g., G)
-- **Mode**: Combination (e.g., G Mixolydian = C major scale with G as tonic)
-
-### Core Analysis Pipeline
-
-1. **Parallel Analysis**: Functional, Modal, and Chromatic analyzers run simultaneously
-2. **Evidence Collection**: Gather cadential, structural, intervallic, and harmonic evidence
-3. **Confidence Calculation**: Weighted evidence scoring with diversity bonus
-4. **Multiple Interpretation Generation**: Primary + alternatives above threshold
-
-### Confidence Thresholds
-
-- **Functional**: 0.4+ (Display threshold)
-- **Modal**: 0.6+ (Display threshold)
-- **Chromatic**: 0.5+ (Display threshold)
-
-## Essential Commands
-
-### Testing
-
-```bash
-# Full comprehensive test suite
-python -m pytest tests/test_comprehensive_multi_layer_validation.py -v
-
-# Edge case behavioral testing
-python -m pytest tests/test_edge_case_behavior.py -v
-
-# Run all tests including edge case validation
-python -m pytest tests/ -v --tb=short
-```
-
-### Development
-
-```bash
-# Setup development environment
-pip install -e .
-pip install -r requirements-dev.txt
-
-# Pre-commit hooks
-pre-commit install
-pre-commit run --all-files
-
-# Code quality
-black src/ tests/ scripts/
-isort src/ tests/ scripts/
-flake8 src/ tests/ scripts/
-mypy src/ --ignore-missing-imports
-```
-
-### Key Scripts
-
-```bash
-# Confidence calibration analysis (essential for debugging)
-python scripts/confidence_calibration_analysis.py
-
-# Regenerate comprehensive test data (when needed)
-python scripts/generate_comprehensive_multi_layer_tests.py
-```
-
-## Critical Files
-
-### Core Analysis Engines
-
-- `src/harmonic_analysis/enhanced_modal_analyzer.py` - Modal analysis engine
-- `src/harmonic_analysis/functional_harmony.py` - Functional harmony engine
-- `src/harmonic_analysis/multiple_interpretation_service.py` - Main orchestrator
-- `src/harmonic_analysis/comprehensive_analysis.py` - Unified analysis entry point
-
-### Testing Infrastructure
-
-- `tests/test_comprehensive_multi_layer_validation.py` - Main validation suite (427 cases)
-- `tests/test_edge_case_behavior.py` - Edge case behavioral validation
-- `scripts/generate_comprehensive_multi_layer_tests.py` - Test data generator
-
-## Current Status
-
-### Test Performance
-
-- **Modal Characteristics**: 56% success rate ✅ (Target: 50%+)
-- **Functional Harmony**: 50%+ success rate ✅ (Target: 50%+)
-- **Overall System**: 28% success rate (limited by modal contextless issues)
-
-### Next Phase Priority
-
-**Modal Contextless Analysis**: 0% success rate on progressions without parent key context (~168 test cases)
-
-## Quality Standards
-
-- All code must pass: black, isort, flake8, mypy
-- Test coverage should remain above 60%
-- Always run comprehensive tests before commits
-- Modal characteristic tests must maintain 60%+ success rate
