+++ docs/README.md
@@ -13,7 +13,7 @@ This directory contains comprehensive documentation for the Harmonic Analysis Py
 ### Specialized Guides
 - **[Troubleshooting](TROUBLESHOOTING.md)** - Common issues, debugging, and advanced usage patterns
 - **[Confidence Calibration](CONFIDENCE_CALIBRATION.md)** - Detailed confidence scoring implementation
-- **[PyPI Setup](PYPI_SETUP.md)** - Guide for automatic PyPI publishing and package distribution
+- **[Suggestion Framework Plan](SUGGESTION_FRAMEWORK_PLAN.md)** - Bidirectional suggestion system architecture
 
 ## Quick Navigation by Topic
 
@@ -37,10 +37,9 @@ This directory contains comprehensive documentation for the Harmonic Analysis Py
 - [Troubleshooting: Performance Optimization](TROUBLESHOOTING.md#performance-optimization)
 - [Troubleshooting: Advanced Usage](TROUBLESHOOTING.md#advanced-usage-patterns)
 
-### Project Setup
-- [PyPI Setup: Account and Token Configuration](PYPI_SETUP.md#prerequisites)
-- [PyPI Setup: Automatic Publishing](PYPI_SETUP.md#how-it-works)
-- [PyPI Setup: Testing and Validation](PYPI_SETUP.md#testing-the-setup)
+### Advanced Topics
+- [Suggestion Framework: Bidirectional Analysis](SUGGESTION_FRAMEWORK_PLAN.md#bidirectional-suggestion-framework)
+- [Architecture: Integration Considerations](ARCHITECTURE.md#integration-considerations)
 
 ## Documentation Maintenance
